chore: refactor logging file locations
