+++ packages/backend/src/clients/EmailClient/EmailClient.ts
@@ -12,7 +12,7 @@ import Mail from 'nodemailer/lib/mailer';
 import { AuthenticationType } from 'nodemailer/lib/smtp-connection';
 import path from 'path';
 import { LightdashConfig } from '../../config/parseConfig';
-import Logger from '../../logger';
+import Logger from '../../logging/logger';
 
 export type AttachmentUrl = {
     path: string;
