+++ packages/backend/src/scheduler/SchedulerJobTimeout.ts
@@ -1,6 +1,6 @@
 import * as Sentry from '@sentry/node';
 import { Job } from 'graphile-worker';
-import Logger from '../logger';
+import Logger from '../logging/logger';
 
 async function timeout(prom: Promise<any>, time: number, exception: Symbol) {
     let timer: NodeJS.Timeout;
