+++ packages/backend/src/scheduler/SchedulerWorkerEventEmitter.ts
@@ -1,7 +1,7 @@
 import * as Sentry from '@sentry/node';
 import EventEmitter from 'events';
 import { WorkerEvents } from 'graphile-worker';
-import Logger from '../logger';
+import Logger from '../logging/logger';
 
 const schedulerWorkerEventEmitter: WorkerEvents = new EventEmitter();
